start of arm subclass
