+++ Arm/Arm/arm.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+namespace sjsu::arm
+{
+
+class Arm
+{
+
+    virtual void Initalize() = 0;
+
+    virtual void HomeArm() = 0;
+
+    virtual void HandleMovement() = 0;
+}; 
+}//sjsu::arm
